feat: add XML serialization to ResourceReferenceList
Split multi-type ResourceReferenceList.cs into one-type-per-file (PNX004).
Sync root csharpier version to 0.29.2 to match c-sharp/ project.
