+++ .config/dotnet-tools.json
@@ -3,7 +3,7 @@
   "isRoot": true,
   "tools": {
     "csharpier": {
-      "version": "0.27.3",
+      "version": "0.29.2",
       "commands": ["dotnet-csharpier"]
     }
   }
