+++ c-sharp/Projects/DblResourceReference.cs
@@ -1,0 +1,6 @@
+namespace Paranext.DataProvider.Projects;
+
+public record DblResourceReference : ResourceReference
+{
+    public string Id { get; init; } = string.Empty;
+}
