+++ c-sharp/Projects/EnhancedResourceReference.cs
@@ -1,0 +1,3 @@
+namespace Paranext.DataProvider.Projects;
+
+public record EnhancedResourceReference : ResourceReference { }
