+++ c-sharp/Projects/ProjectReference.cs
@@ -1,0 +1,6 @@
+namespace Paranext.DataProvider.Projects;
+
+public record ProjectReference : ResourceReference
+{
+    public string Id { get; init; } = string.Empty;
+}
