+++ c-sharp/Projects/SourceLanguageResourceReference.cs
@@ -1,0 +1,3 @@
+namespace Paranext.DataProvider.Projects;
+
+public record SourceLanguageResourceReference : ResourceReference { }
