+++ c-sharp/Projects/XmlResourceReference.cs
@@ -1,0 +1,3 @@
+namespace Paranext.DataProvider.Projects;
+
+public record XmlResourceReference : ResourceReference { }
